normie and shel lvl done
