+++ incs/mini_header.h
@@ -6,7 +6,7 @@
 /*   By: dpaes-so <dpaes-so@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/11 16:55:53 by dpaes-so          #+#    #+#             */
-/*   Updated: 2025/05/05 13:01:28 by dpaes-so         ###   ########.fr       */
+/*   Updated: 2025/05/05 15:51:38 by dpaes-so         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -116,6 +116,9 @@ void					middle_child(t_mini *mini, t_cmd cmds);
 void					solo_child(t_mini *mini, t_cmd cmds);
 int						here_doc(t_pipe pipex, t_cmd *cmds);
 void					exit_childprocess_exec(t_mini *mini);
+void					run_tree(t_mini *mini, t_tree *ast, int f);
+void					wait_child(t_mini *mini);
+void					cmd_exit(char *exec, t_mini *mini);
 
 //----------------------------PARSING ! ! ! -------------------------------
